+++ CHANGELOG.md
@@ -33,7 +33,7 @@ _TBD_
 - [x] Added new methods to `GatewayAccount`: `getAdditionalFilters`, `setAdditionalFilters`
 - [x] Added new property to `GatewayAccount`: `digitalWallets`
 - [x] Added new methods to `GatewayAccount`: `setDigitalWallets`, `getDigitalWallets`
-- [x] Added new property to `Invoice`: `transactions`
+- [x] Added new properties to `Invoice`: `transactions`, `poNumber`, `notes`
 - [x] Added new property to `Subscription`: `recentInvoice`
 - [x] Added `patch` method to `TransactionService`